+++ test/src/SanitizerTest.php
@@ -165,7 +165,7 @@ class SanitizerTest extends \PHPUnit_Framework_TestCase {
      * @covers Sanitor\Sanitizer::filterRequest
      */
     public function testFilterPost() {
-       $curl = curl_init('localhost:8080/index.php');
+       $curl = curl_init('localhost:8080/post.php');
         curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($curl, CURLOPT_POST, true);
         curl_setopt($curl, CURLOPT_POSTFIELDS, array('email' => 'mail@benedict\roeser.de'));
@@ -213,7 +213,7 @@ class SanitizerTest extends \PHPUnit_Framework_TestCase {
      * @covers Sanitor\Sanitizer::filterInput
      */
     public function testFilterGet() {
-        $curl = curl_init('localhost:8080/index.php?email='.  rawurlencode('mail@benedict\roeser.de'));
+        $curl = curl_init('localhost:8080/get.php?email='.  rawurlencode('mail@benedict\roeser.de'));
         curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
         $output = curl_exec($curl);
         curl_close($curl);
@@ -278,14 +278,32 @@ class SanitizerTest extends \PHPUnit_Framework_TestCase {
 
     /**
      * @covers Sanitor\Sanitizer::filterEnv
-     * @covers Sanitor\Sanitizer::filterInput
-     * @todo   Implement testFilterEnv().
+     * @covers Sanitor\Sanitizer::checkSanitizedValue
      */
     public function testFilterEnv() {
-        // Remove the following lines when you implement this test.
-        $this->markTestIncomplete(
-                'This test has not been implemented yet.'
-        );
+        $curl = curl_init('localhost:8080/env.php');
+        curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
+        $output = curl_exec($curl);
+        curl_close($curl);
+        $output = json_decode($output);
+        if(!is_array($output)) {
+            throw new \Exception('Invalid JSON data');
+        }
+        $this->assertEquals('BAR', array_shift($output));
+        $this->assertNull(array_shift($output));
+        $this->assertEquals('EXCEPTION', array_shift($output));
+
+        // The rest of this method just ensures that code coverage reports are
+        // correct; This is extremely ugly, but it works
+        putenv('BAZ=DING');
+        $this->object->filterEnv('BAZ');
+        $this->object->filterEnv('username');
+        try {
+            $this->object->filterEnv(42);
+        } catch (\Exception $ex) {
+            return;
+        }
+        $this->fail('An Exception should be thrown');
     }
 
     /**